+++ selfdrive/car/honda/interface.py
@@ -55,9 +55,17 @@ class CarInterface(CarInterfaceBase):
     if any(0x33DA in f for f in fingerprint.values()):
       ret.flags |= HondaFlags.BOSCH_EXT_HUD.value
 
-    # Accord 1.5T CVT has different gearbox message
-    if candidate == CAR.ACCORD and 0x191 in fingerprint[1]:
-      ret.transmissionType = TransmissionType.cvt
+    # Detect transmission type for Accord
+    if candidate in [CAR.ACCORD, CAR.ACCORDH]:
+      # Accord 1.5
+      if 0x191 in fingerprint[1]:
+        ret.transmissionType = TransmissionType.cvt
+      # Accord 2.0 and Hybrid
+      elif 0x1a3 in fingerprint[1]:
+        ret.transmissionType = TransmissionType.automatic
+      else:
+        ret.transmissionType = TransmissionType.manual
+
 
     # Certain Hondas have an extra steering sensor at the bottom of the steering rack,
     # which improves controls quality as it removes the steering column torsion from feedback.